Best PE results thus far
Did not change much with the network structure of the GAN. One thing that does seem to add more spread is by decreasing the Gaussian Dropout to .01. What made a massive difference was
changing the activation function of the CNN PE network.

Changed activation function of mass ratio param to be sigmoid throughout.